refactor: created a separate file for the centralizer class
